backend/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
httpx==0.28.1
pydantic==2.12.5
redis[hiredis]==7.4.0
python-dotenv==1.2.2
